Extract types from documentUnit.ts
RISDEV-8322
