+++ frontend/src/components/HandoverDuplicateCheckView.vue
@@ -3,7 +3,7 @@ import Button from "primevue/button"
 import { computed } from "vue"
 import DecisionSummary from "@/components/DecisionSummary.vue"
 import { useInternalUser } from "@/composables/useInternalUser"
-import { DuplicateRelation } from "@/domain/documentUnit"
+import { DuplicateRelation } from "@/domain/managementData"
 import DateUtil from "@/utils/dateUtil"
 import IconCheck from "~icons/ic/baseline-check"
 import IconErrorOutline from "~icons/ic/baseline-error-outline"
