+++ frontend/src/components/category-import/CategoryImport.vue
@@ -9,8 +9,7 @@ import DecisionSummary from "@/components/DecisionSummary.vue"
 import InputField from "@/components/input/InputField.vue"
 import { useValidationStore } from "@/composables/useValidationStore"
 import ActiveCitation from "@/domain/activeCitation"
-import { Decision } from "@/domain/decision"
-import { longTextLabels, shortTextLabels } from "@/domain/documentUnit"
+import { Decision, longTextLabels, shortTextLabels } from "@/domain/decision"
 import NormReference from "@/domain/normReference"
 import ParticipatingJudge from "@/domain/participatingJudge"
 import Reference from "@/domain/reference"
