+++ frontend/src/components/inbox/PendingHandover.vue
@@ -6,7 +6,8 @@ import { InfoStatus } from "@/components/enumInfoStatus"
 import InfoModal from "@/components/InfoModal.vue"
 import { Page } from "@/components/Pagination.vue"
 import { Query } from "@/composables/useQueryFromRoute"
-import { DocumentUnitSearchParameter, InboxStatus } from "@/domain/documentUnit"
+import { InboxStatus } from "@/domain/decision"
+import { DocumentUnitSearchParameter } from "@/domain/documentUnit"
 import DocumentUnitListEntry from "@/domain/documentUnitListEntry"
 import service from "@/services/documentUnitService"
 import { ResponseError } from "@/services/httpClient"
