+++ frontend/src/components/inbox/shared/InboxList.vue
@@ -11,7 +11,7 @@ import InputErrorMessages from "@/components/InputErrorMessages.vue"
 import Pagination, { Page } from "@/components/Pagination.vue"
 import PopupModal from "@/components/PopupModal.vue"
 import { useStatusBadge } from "@/composables/useStatusBadge"
-import { InboxStatus } from "@/domain/documentUnit"
+import { InboxStatus } from "@/domain/decision"
 import DocumentUnitListEntry from "@/domain/documentUnitListEntry"
 import { ResponseError } from "@/services/httpClient"
 import IconAttachedFile from "~icons/ic/baseline-attach-file"
