+++ frontend/src/components/preview/DocumentUnitPreview.vue
@@ -15,8 +15,9 @@ import PreviewLongTexts from "@/components/preview/PreviewLongTexts.vue"
 import PreviewNote from "@/components/preview/PreviewNote.vue"
 import PreviewProceedingDecisions from "@/components/preview/PreviewProceedingDecisions.vue"
 import PreviewShortTexts from "@/components/preview/PreviewShortTexts.vue"
-import { Decision } from "@/domain/decision"
-import { ContentRelatedIndexing, Kind, LongTexts } from "@/domain/documentUnit"
+import { ContentRelatedIndexing } from "@/domain/contentRelatedIndexing"
+import { Decision, LongTexts } from "@/domain/decision"
+import { Kind } from "@/domain/documentUnit"
 import EnsuingDecision from "@/domain/ensuingDecision"
 import PreviousDecision from "@/domain/previousDecision"
 import Reference from "@/domain/reference"
