+++ frontend/src/components/texts/DocumentUnitLongTexts.vue
@@ -3,8 +3,7 @@ import { storeToRefs } from "pinia"
 import { type Component, computed, Ref } from "vue"
 import TextEditorCategory from "@/components/texts/TextEditorCategory.vue"
 import { useValidBorderNumberLinks } from "@/composables/useValidBorderNumberLinks"
-import { Decision } from "@/domain/decision"
-import { longTextLabels } from "@/domain/documentUnit"
+import { Decision, longTextLabels } from "@/domain/decision"
 import { useDocumentUnitStore } from "@/stores/documentUnitStore"
 import TextEditorUtil from "@/utils/textEditorUtil"
