+++ frontend/src/components/texts/DocumentUnitShortTexts.vue
@@ -4,8 +4,7 @@ import { Component, Ref, computed } from "vue"
 import TextEditorCategory from "@/components/texts/TextEditorCategory.vue"
 import TextInputCategory from "@/components/texts/TextInputCategory.vue"
 import { useValidBorderNumberLinks } from "@/composables/useValidBorderNumberLinks"
-import { Decision } from "@/domain/decision"
-import { shortTextLabels } from "@/domain/documentUnit"
+import { Decision, shortTextLabels } from "@/domain/decision"
 import { useDocumentUnitStore } from "@/stores/documentUnitStore"
 import TextEditorUtil from "@/utils/textEditorUtil"
