+++ frontend/src/domain/source.ts
@@ -1,0 +1,16 @@
+import Reference from "@/domain/reference"
+
+export enum SourceValue {
+  UnaufgefordertesOriginal = "O",
+  AngefordertesOriginal = "A",
+  Zeitschrift = "Z",
+  Email = "E",
+  LaenderEuGH = "L",
+  Sonstige = "S",
+}
+
+export type Source = {
+  value?: SourceValue
+  reference?: Reference
+  sourceRawValue?: string
+}
