+++ frontend/test/components/handoverDocument.spec.ts
@@ -5,10 +5,10 @@ import { beforeEach } from "vitest"
 import { createRouter, createWebHistory } from "vue-router"
 import HandoverDocumentationUnitView from "@/components/HandoverDocumentationUnitView.vue"
 import { Decision } from "@/domain/decision"
-import { DuplicateRelationStatus } from "@/domain/documentUnit"
 import { Env } from "@/domain/env"
 import { EventRecordType, HandoverMail, Preview } from "@/domain/eventRecord"
 import LegalForce from "@/domain/legalForce"
+import { DuplicateRelationStatus } from "@/domain/managementData"
 import NormReference from "@/domain/normReference"
 import SingleNorm from "@/domain/singleNorm"
 import featureToggleService from "@/services/featureToggleService"
