+++ frontend/test/components/inbox/shared/inboxList.spec.ts
@@ -3,7 +3,7 @@ import { render, screen } from "@testing-library/vue"
 import { createRouter, createWebHistory } from "vue-router"
 import InboxList from "@/components/inbox/shared/InboxList.vue"
 import { Page } from "@/components/Pagination.vue"
-import { InboxStatus } from "@/domain/documentUnit"
+import { InboxStatus } from "@/domain/decision"
 import DocumentUnitListEntry from "@/domain/documentUnitListEntry"
 import { PublicationState } from "@/domain/publicationStatus"
 import routes from "~/test-helper/routes"
