+++ frontend/test/components/input/textEditor.spec.ts
@@ -6,7 +6,7 @@ import { flushPromises } from "@vue/test-utils"
 import { createRouter, createWebHistory } from "vue-router"
 import TextEditor from "@/components/input/TextEditor.vue"
 import { TextAreaInputAttributes } from "@/components/input/types"
-import { longTextLabels } from "@/domain/documentUnit"
+import { longTextLabels } from "@/domain/decision"
 import { useFeatureToggleServiceMock } from "~/test-helper/useFeatureToggleServiceMock"
 
 useFeatureToggleServiceMock()
