+++ frontend/test/components/management-data/assignDocumentationOffice.spec.ts
@@ -7,7 +7,7 @@ import { describe, it } from "vitest"
 import { createRouter, createWebHistory } from "vue-router"
 import DocumentUnitManagementData from "@/components/management-data/DocumentUnitManagementData.vue"
 import { Decision } from "@/domain/decision"
-import { ManagementData } from "@/domain/documentUnit"
+import { ManagementData } from "@/domain/managementData"
 import errorMessages from "@/i18n/errors.json"
 import DocumentUnitHistoryLogService from "@/services/documentUnitHistoryLogService"
 import documentUnitService from "@/services/documentUnitService"
