+++ frontend/test/components/preview/previewContentRelatedIndexing.spec.ts
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/vue"
 import { previewLayoutInjectionKey } from "@/components/preview/constants"
 import PreviewContentRelatedIndexing from "@/components/preview/PreviewContentRelatedIndexing.vue"
 import ActiveCitation from "@/domain/activeCitation"
-import { ContentRelatedIndexing } from "@/domain/documentUnit"
+import { ContentRelatedIndexing } from "@/domain/contentRelatedIndexing"
 import NormReference from "@/domain/normReference"
 import SingleNorm from "@/domain/singleNorm"
