+++ frontend/test/components/preview/previewPendingProceeding.spec.ts
@@ -2,8 +2,8 @@ import { createTestingPinia } from "@pinia/testing"
 import { render, screen } from "@testing-library/vue"
 import { beforeAll } from "vitest"
 import PendingProceedingPreview from "@/components/preview/PendingProceedingPreview.vue"
-import { SourceValue } from "@/domain/documentUnit"
 import PendingProceeding from "@/domain/pendingProceeding"
+import { SourceValue } from "@/domain/source"
 import { onSearchShortcutDirective } from "@/utils/onSearchShortcutDirective"
 import { useFeatureToggleServiceMock } from "~/test-helper/useFeatureToggleServiceMock"
